refactor: replace SetMemberWageModalCopy with SetMemberWageModal and update related imports for consistency
