+++ app/src/components/sections/DashboardView/SetMemberWageModal.vue
