+++ app/src/components/sections/DashboardView/SetMemberWageStandardStep.vue
@@ -78,7 +78,7 @@
 <script setup lang="ts">
 import * as z from 'zod'
 import { NETWORK } from '@/constant'
-import type { WageWithForm } from './SetMemberWageModalCopy.vue'
+import type { WageWithForm } from './SetMemberWageModal.vue'
 
 const wageData = defineModel<WageWithForm>('wageData', { required: true })
